Moved mocks to correct dir
Related tasks APPLINK-30588 APPLINK-30974
